% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_tune.R
\name{random_tune}
\alias{random_tune}
\title{Randomly select a tune}
\usage{
random_tune(artist = "Amerie", data = billboard::wiki_hot_100s)
}
\arguments{
\item{artist}{An artist/group to select a song from (default=Amerie)}

\item{data}{The dataset to use (default= billboard::wiki_hot_100s)}
}
\value{
A random song is printed to the console
}
\description{
Randomly select a tune
}
\examples{
random_tune()
random_tune(artist = "Ray Charles")
random_tune("Salt-n-Pepa")

}
